+++ src/structural_graph/ast.rs
@@ -8,15 +8,15 @@ pub enum EntryType {
     ControlReg,
 }
 
-#[derive(PartialEq, Eq, Debug)]
+#[derive(PartialEq, Debug)]
 pub struct Entry {
     pub entry_type: EntryType,
     pub name: Symbol,
-    pub adjacency_list: Vec<(Symbol, usize)>,
+    pub adjacency_list: Vec<(Symbol, f64)>,
 }
 
 impl Entry {
-    pub fn new(entry_type: EntryType, name: Symbol, adjacency_list: Vec<(Symbol, usize)>) -> Entry {
+    pub fn new(entry_type: EntryType, name: Symbol, adjacency_list: Vec<(Symbol, f64)>) -> Entry {
         Entry {
             entry_type,
             name,
